refactor(timeline): コンポーネント分割
